dwh-project/requirements.txt
boto3==1.16.51
psycopg2-binary==2.8.6
